Check the necessity of upgrade; Download updates with wget instead of git; Minor fixes

+++ update_server/latest.sh
@@ -2,6 +2,7 @@
 
 APPDIR="/home/a3/a3-metro-train"
 TEMPDIR="/tmp/metro-update"
+TEMPFILE="${TEMPDIR}.zip"
 
 die() {
     echo 'Error executing command, exiting'
@@ -9,20 +10,27 @@ die() {
 }
 
 # Check current version
-# TODO
+LOCALVERSION=`cat ${APPDIR}/build.txt`;
+
+if test $LOCALVERSION = "2017.02.12"; then
+    echo "Update is not required";
+    exit 1
+fi
 
 # Cleanup previous installation
 rm -rf $TEMPDIR || true
+rm -rf $TEMPFILE || true
 mkdir -p $TEMPDIR || die
 
 # Download latest project files and dependencies
-git clone --branch release https://github.com/agentyzmin/metro-train-screens.git $TEMPDIR || die
+wget https://metro-update.veikus.com/download/build_2017.02.12.zip -O $TEMPFILE || die
+unzip $TEMPFILE -d $TEMPDIR || die
 npm i --production --prefix $TEMPDIR || die
 
 # Update file permissions
-chmod +x "${TEMPDIR}/scripts/*.sh" || die
+chmod -v +x ${TEMPDIR}/scripts/*.sh || die
 
 # Replace original directory with updated files
-chown -R a3:a3 $TEMPDIR "${TEMPDIR}/*" "${TEMPDIR}/.[!.]*" || die
+chown -R a3:a3 $TEMPDIR ${TEMPDIR}/* ${TEMPDIR}/.[!.]* || die
 rm -rf $APPDIR || die
 mv $TEMPDIR $APPDIR || die
